- Added the following config options for auto_menu: wait_for_min_players, wait_for_max_players, and wait_for_time - Issue 381 and issue 373 are almost finished, but still have some issues. - Add incomplete/experimental DirectIPServer project. - Removed the logging_str config option and references.
git-svn-id: https://bwapi.googlecode.com/svn/trunk@3709 273ecf2c-f24b-0410-b6a7-e9298bb4873f

+++ bwapi/BWAPI/Source/Recording.cpp
@@ -35,7 +35,7 @@ bool StartVideoRecording(const char *pszFilename, int width, int height, BITMAPI
   aisinfo.dwScale               = 1;
   aisinfo.dwRate                = 30;
   aisinfo.dwSuggestedBufferSize = width * height;
-  aisinfo.dwQuality             = ~0;
+  aisinfo.dwQuality             = (DWORD)~0;
   aisinfo.dwSampleSize          = width * height;
   SetRect(&aisinfo.rcFrame, 0, 0, width, height);
